app/src/main/java/com/jusixs/ndrs/data/repository/TrainingRepository.java
package com.jusixs.ndrs.data.repository;

import com.jusixs.ndrs.data.model.TrainingSession;
import java.util.Collections;
import java.util.List;

/**
 * Repository class for managing training-related data and operations.
 */
public class TrainingRepository {

    /**
     * Retrieves training sessions filtered by role and region.
     *
     * @param role   User's role to filter sessions.
     * @param region Region to filter sessions.
     * @return List of training sessions matching the role and region criteria.
     */
    public List<TrainingSession> getTrainingSessionsByRoleAndRegion(String role, String region) {
        return Collections.emptyList(); // Replace with actual fetch logic
    }

    /**
     * Registers a user for a specific training session.
     *
     * @param sessionId Unique identifier of the session.
     * @param userId    Unique identifier of the user.
     * @return true if registration is successful, false otherwise.
     */
    public boolean registerForSession(String sessionId, String userId) {
        return true; // Replace with actual registration logic
    }

    /**
     * Retrieves training materials for a given session.
     *
     * @param sessionId Unique identifier of the session.
     * @return List of training materials related to the session.
     */
    public List<String> getTrainingMaterials(String sessionId) {
        return Collections.emptyList(); // Replace with actual fetch logic
    }

    /**
     * Marks attendance for an offline session.
     *
     * @param sessionId Unique identifier of the session.
     * @param userId    Unique identifier of the user.
     * @return true if attendance is marked successfully, false otherwise.
     */
    public boolean markAttendanceOffline(String sessionId, String userId) {
        return true; // Replace with actual attendance logic
    }

    /**
     * Logs attendance for an online session.
     *
     * @param sessionId Unique identifier of the session.
     * @param userId    Unique identifier of the user.
     * @return true if attendance is logged successfully, false otherwise.
     */
    public boolean logAttendanceOnline(String sessionId, String userId) {
        return true; // Replace with actual logging logic
    }

    /**
     * Generates a certificate for a user completing a session.
     *
     * @param sessionId Unique identifier of the session.
     * @param userId    Unique identifier of the user.
     * @return true if certificate generation is successful, false otherwise.
     */
    public boolean generateCertificate(String sessionId, String userId) {
        return true; // Replace with actual generation logic
    }

    /**
     * Stores a certificate in the user's profile.
     *
     * @param userId        Unique identifier of the user.
     * @param certificateId Unique identifier of the certificate.
     */
    public void storeCertificateInProfile(String userId, String certificateId) {
        // Implement logic to store certificate in user profile
    }

    /**
     * Submits feedback for a session.
     *
     * @param sessionId Unique identifier of the session.
     * @param userId    Unique identifier of the user.
     * @param feedback  Feedback content provided by the user.
     * @return true if feedback submission is successful, false otherwise.
     */
    public boolean submitFeedback(String sessionId, String userId, String feedback) {
        return true; // Replace with actual feedback submission logic
    }

    /**
     * Retrieves notifications for the user.
     *
     * @param userId Unique identifier of the user.
     * @return List of notifications for the user.
     */
    public List<String> getNotifications(String userId) {
        return Collections.emptyList(); // Replace with actual fetch logic
    }
}
